Refactor code #JC19-407
